fix(tests): Fix tests across many files
why you so broke synchronous pattern lab?
