+++ test/index_tests.js
@@ -3,7 +3,8 @@ const defaultConfig = require('../patternlab-config.json');
 
 const entry = require('../core/index');
 
-tap.test('getDefaultConfig - should return the default config object', function(test) {
+tap.test('getDefaultConfig - should return the default config object', function (test) {
+  console.log(1)
   const requestedConfig = entry.getDefaultConfig();
   test.type(requestedConfig, 'object');
   test.equals(requestedConfig, defaultConfig);
