+++ test/pseudopattern_hunter_tests.js
@@ -8,6 +8,7 @@ var pph = require('../core/lib/pseudopattern_hunter');
 var pa = require('../core/lib/pattern_assembler');
 var Pattern = require('../core/lib/object_factory').Pattern;
 var PatternGraph = require('../core/lib/pattern_graph').PatternGraph;
+const addPattern = require('../core/lib/addPattern');
 
 var config = require('./util/patternlab-config.json');
 var engineLoader = require('../core/lib/pattern_engines');
@@ -47,7 +48,7 @@ tap.test('pseudpattern found and added as a pattern', function (test) {
   var pl = stubPatternlab();
 
   var atomPattern = pattern_assembler.load_pattern_iterative('00-test/03-styled-atom.mustache', pl);
-  pattern_assembler.addPattern(atomPattern, pl);
+  addPattern(atomPattern, pl);
 
   //act
   var patternCountBefore = pl.patterns.length;
@@ -65,7 +66,7 @@ tap.test('pseudpattern does not pollute base pattern data', function (test) {
   var pl = stubPatternlab();
 
   var atomPattern = pattern_assembler.load_pattern_iterative('00-test/03-styled-atom.mustache', pl);
-  pattern_assembler.addPattern(atomPattern, pl);
+  addPattern(atomPattern, pl);
 
   //act
   var patternCountBefore = pl.patterns.length;
@@ -92,8 +93,8 @@ tap.test('pseudpattern variant includes stylePartials and parameteredPartials', 
   pseudoPattern.stylePartials = pseudoPattern.findPartialsWithStyleModifiers(pseudoPattern);
   pseudoPattern.parameteredPartials = pseudoPattern.findPartialsWithPatternParameters(pseudoPattern);
 
-  pattern_assembler.addPattern(atomPattern, pl);
-  pattern_assembler.addPattern(pseudoPattern, pl);
+  addPattern(atomPattern, pl);
+  addPattern(pseudoPattern, pl);
 
   //act
   return pph.find_pseudopatterns(pseudoPattern, pl).then(() => {
